fix: del/patch with json body

+++ app/controllers/leads_controller.py
@@ -1,16 +1,12 @@
 import re
 from flask import jsonify, request, current_app
 from sqlalchemy.exc import IntegrityError
-from werkzeug.exceptions import NotFound
 from sqlalchemy.exc import NoResultFound
+from werkzeug.exceptions import NotFound
 
-from app.models.exceptions_model import InvalidEmail, InvalidPhone
+from app.models.exceptions_model import InvalidEmail, InvalidKeys, InvalidPhone, InvalidValueType
 from app.models.leads_model import Leads_Model
 from app.services.helpers import data_to_patch, integrity_err_returns
-
-    # print('*'*100)
-    # print(query.visits)
-    # print('*'*100)
 
 def create_leads():
     data = request.get_json()
@@ -18,6 +14,9 @@ def create_leads():
     email_regex = "(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
     
     try:
+        if type(data['name']) != str or type(data['email']) != str or type(data['phone']) != str:
+            raise InvalidValueType
+
         if not re.fullmatch(phone_regex, data['phone']):
             raise InvalidPhone
         
@@ -30,6 +29,9 @@ def create_leads():
         current_app.db.session.commit()
 
         return jsonify(new_leads), 201
+    
+    except InvalidValueType as ivt:
+        return ivt.description, ivt.code
 
     except IntegrityError as i:
         return integrity_err_returns(i, data)
@@ -42,7 +44,7 @@ def create_leads():
 
 def get_leads():
     try:
-        leads = Leads_Model.query.all()
+        leads = Leads_Model.query.order_by(Leads_Model.visits.desc()).all()
 
         if len(leads) == 0:
             raise NotFound
@@ -53,10 +55,18 @@ def get_leads():
     except NotFound:
         return {'error': 'No data found on database'}, 404
 
-def patch_leads(email):
+def patch_leads():
+    data = request.get_json()
+    email = data['email']
     email_regex = "(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
 
     try:
+        if len(data.keys()) > 1:
+            raise InvalidKeys
+        
+        if type(email) != str:
+            raise InvalidValueType
+
         if not re.fullmatch(email_regex, email):
             raise InvalidEmail
 
@@ -72,13 +82,30 @@ def patch_leads(email):
     except NoResultFound:
         return {'error': f'{email} does not exist on database'}, 404
 
+    except InvalidKeys as ik:
+        return ik.description, ik.code
+    
+    except InvalidValueType as ivt:
+        return ivt.description, ivt.code
+
     except InvalidEmail as ie:
         return ie.email_err_description(email), ie.code
 
-def delete_leads(email):
-    # data = request.get_json()
+def delete_leads():
+    data = request.get_json()
+    email = data['email']
+    email_regex = "(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
 
     try:
+        if len(data.keys()) > 1:
+            raise InvalidKeys
+        
+        if type(email) != str:
+            raise InvalidValueType
+
+        if not re.fullmatch(email_regex, email):
+            raise InvalidEmail
+
         query = Leads_Model.query.filter_by(email=email).one()
 
         current_app.db.session.delete(query)
@@ -88,3 +115,12 @@ def delete_leads(email):
 
     except NoResultFound:
         return {'error': f'{email} does not exist on database'}, 404
+    
+    except InvalidKeys as ik:
+        return ik.description, ik.code
+    
+    except InvalidValueType as ivt:
+        return ivt.description, ivt.code
+
+    except InvalidEmail as ie:
+        return ie.email_err_description(email), ie.code
